requirements.txt
Flask==2.3.0
flask-cors==4.0.0
python-chess==1.999
PyJWT==2.8.0
Werkzeug==2.3.0
python-socketio==5.9.0
python-engineio==4.7.1